Make cloud status check interval configurable.
We run a large number of build hosts and spot fleet instances in an AWS account and the
default setting of 10 seconds for the CloudNanny checker ran us into API limits with AWS.

This change allows us to set the check interval from 10 seconds to 60 seconds thus keeping
us under the API limits.

This patch adds a new configuration option in the "advanced" section of the EC2 fleet configuration which
controls the interval at which the CloudNanny checker is run.